add visual components
